+++ packages/web/src/common/store/account/sagas.js
@@ -23,7 +23,7 @@ import {
 
 import { waitForBackendSetup } from 'common/store/backend/sagas'
 import { retrieveCollections } from 'common/store/cache/collections/utils'
-import { updateProfileAsync } from 'pages/profile-page/sagas'
+import { updateProfileAsync } from 'common/store/profile/sagas'
 import { recordIP } from 'services/audius-backend/RecordIP'
 import { createUserBankIfNeeded } from 'services/audius-backend/waudio'
 import { fingerprintClient } from 'services/fingerprint'
