+++ packages/web/src/store/sagas.ts
@@ -21,6 +21,7 @@ import explorePageSagas from 'common/store/pages/explore/sagas'
 import feedPageSagas from 'common/store/pages/feed/sagas'
 import signOnSaga from 'common/store/pages/signon/sagas'
 import playerSagas from 'common/store/player/sagas'
+import profileSagas from 'common/store/profile/sagas'
 import mobileQueueSagas from 'common/store/queue/mobileSagas'
 import queueSagas from 'common/store/queue/sagas'
 import reachabilitySagas from 'common/store/reachability/sagas'
@@ -48,7 +49,6 @@ import followersPageSagas from 'pages/followers-page/sagas'
 import followingPageSagas from 'pages/following-page/sagas'
 import historySagas from 'pages/history-page/store/sagas'
 import mutualsPageSagas from 'pages/mutuals-page/sagas'
-import profileSagas from 'pages/profile-page/sagas'
 import remixesSagas from 'pages/remixes-page/store/sagas'
 import repostPageSagas from 'pages/reposts-page/sagas'
 import savedSagas from 'pages/saved-page/store/sagas'
